Only allow execution from vcl_error
While execution in, for instance, vcl_deliver, might work "most of the
time", it is inherently unsafe because we are ripping apart shared
resources. This commit ensures, quite thoroughly, that only access from
vcl_error is allowed.

Since we can't do it at compile-time, our only option is to verify at
run-time. We try to throw a printf explaining it too, before we fail by
assert.